🧪 test: update tests to match style changes
- move input tests into their own test module
- add new test module for new choice field
- simplify test names--removed redundant "test" in file name
